modified project structure and cmdline tools for project_df; setup a messaging context manager
