--- spacemake/cmdline/__init__.py
@@ -1,2 +1,0 @@
-from ._config import ConfigFile
-from ._project_df import ProjectDF
