--- spacemake/config/config.yaml
@@ -1,69 +1,0 @@
-microscopy_out: '/data/rajewsky/slideseq_microscopy/qc/'
-
-external_bin:
-    dropseq_tools: '/data/rajewsky/shared_bins/Drop-seq_tools-2.4.0'
-    picard_tools: '/data/rajewsky/shared_bins/picard-tools-2.9.0/picard.jar'
-
-puck_data:
-    barcode_file: 'predictions_ml.csv'
-    root: 'puck_data'
-
-pucks:
-    default:
-        width_um: 3000
-        spot_diameter_um: 10
-    visium:
-        barcodes: 'puck_data/visium_barcode_positions.csv'
-        width_um: 6500
-        spot_diameter_um: 55
-
-run_modes:
-    default:
-        n_beads: 100000
-        umi_cutoff: [100, 300]
-        clean_dge: True
-        plot_bead_size: 0.1
-        detect_tissue: False
-        polyA_adapter_trimming: True
-        count_intronic_reads: True
-        count_mm_reads: True
-        mesh_data: False
-        mesh_spot_diameter_um: 55
-        mesh_spot_distance_um: 100
-    visium:
-        n_beads: 10000
-        umi_cutoff: [1000]
-        clean_dge: False
-        plot_bead_size: 1
-        detect_tissue: True
-        polyA_adapter_trimming: False
-        count_intronic_reads: False
-        count_mm_reads: True
-    slideseq:
-        expected_n_beads: 100000
-        umi_cutoff: [50]
-        clean_dge: False
-        plot_bead_size: 0.1
-        detect_tissue: False
-
-adapters:
-    smart: 'AAGCAGTGGTATCAACGCAGAGTGAATGGG'
-    optical_primer: 'GAATCACGATACGTACACCA'
-
-barcode_flavors:
-    default:
-        cell: "r1[0:12]"
-        UMI: "r1[12:20]"
-        bam_tags: "CR:{cell},CB:{cell},MI:{UMI},RG:{assigned}"
-    slideseq_14bc:
-        cell: "r1[0:14]"
-        UMI: "r1[14:23]"
-        bam_tags: "CR:{cell},XC:{cell},XM:{UMI},RG:{assigned}"
-    slideseq_15bc:
-        cell: "r1[0:14]"
-        UMI: "r1[15:23]"
-        bam_tags: "CR:{cell},XC:{cell},XM:{UMI},RG:{assigned}"
-    visium:
-        cell: "r1[0:16]"
-        UMI: "r1[16:28]"
-        bam_tags: "CR:{cell},CB:{cell},MI:{UMI},RG:{assigned}"
