--- spacemake/spatial/__init__.py
@@ -1,1 +1,0 @@
-from ._spatial import create_meshed_adata, detect_tissue
